+++ valkey_glide_cluster.c
@@ -1002,6 +1002,10 @@ CLIENT_PAUSE_METHOD_IMPL(ValkeyGlideCluster)
 CLIENT_UNPAUSE_METHOD_IMPL(ValkeyGlideCluster)
 /* }}} */
 
+/* {{{ proto array ValkeyGlideCluster::clientTrackingInfo([mixed route]) */
+CLIENT_TRACKING_INFO_METHOD_IMPL(ValkeyGlideCluster)
+/* }}} */
+
 /* {{{ proto string ValkeyGlideCluster::memoryDoctor() */
 MEMORY_DOCTOR_METHOD_IMPL(ValkeyGlideCluster)
 /* }}} */