+++ valkey_z_php_methods.c
@@ -843,6 +843,10 @@ CLIENT_PAUSE_METHOD_IMPL(ValkeyGlide)
 CLIENT_UNPAUSE_METHOD_IMPL(ValkeyGlide)
 /* }}} */
 
+/* {{{ proto array ValkeyGlide::clientTrackingInfo() */
+CLIENT_TRACKING_INFO_METHOD_IMPL(ValkeyGlide)
+/* }}} */
+
 /* {{{ proto string ValkeyGlide::memoryDoctor() */
 MEMORY_DOCTOR_METHOD_IMPL(ValkeyGlide)
 /* }}} */